+++ Tasks/UnityActivateLicense/UnityActivateLicenseV1/post-unity-activate-license.ts
@@ -32,7 +32,6 @@ function run() {
         const unityEditorsPath = UnityPathTools.getUnityEditorsPath(
             tl.getInput(unityEditorsPathModeInputVariableName, true)!,
             tl.getInput(customUnityEditorsPathInputVariableName));
-        const unityProjectPath = tl.getPathInput(unityProjectPathInputVariableName) ?? '';
 
         let unityVersion: UnityVersionInfoResult;
         if (versionSelectionMode === 'specify') {
@@ -67,7 +66,7 @@ function run() {
                 .arg('-returnlicense')
                 .arg('-logfile').arg(logFilePath)
                 .arg('-quit');
-                
+
             const result = unityCmd.execSync();
 
             // Unity process has finished. Set task result.